added PicturesUpload Component
